better destroy

+++ system/scripts/xvirt/xvirt-destroy.sh
@@ -2,9 +2,19 @@
 
 set -e
 set -u
-set -x
+#set -x
 
 source vm.profile
 
-virsh destroy ${DT_VM_HOSTNAME}
-virsh undefine ${DT_VM_HOSTNAME} --remove-all-storage
+state=$(virsh domstate ${DT_VM_HOSTNAME} 2>/dev/null || echo "NONEXIST")
+state=`echo "${state}" | xargs`
+
+if [ "${state}" = "NONEXIST" ]; then
+	echo "Vm '${DT_VM_HOSTNAME}' doesn't exist."
+else
+	if [ "${state}" = "running" ]; then
+		virsh destroy ${DT_VM_HOSTNAME}
+	fi
+
+	virsh undefine ${DT_VM_HOSTNAME} --remove-all-storage
+fi
